refactor: change role based access control

+++ internal/modules/users/commands/delete_by_id.go
@@ -3,6 +3,7 @@ package commands
 import (
 	"github.com/italoservio/serviosoftwareusers/internal/modules/users/repos"
 	"github.com/italoservio/serviosoftwareusers/pkg/exception"
+	"github.com/italoservio/serviosoftwareusers/pkg/jwt"
 )
 
 type DeleteUserByIDCmd struct {
@@ -10,7 +11,8 @@ type DeleteUserByIDCmd struct {
 }
 
 type DeleteUserByIDCmdInput struct {
-	ID string `validate:"required,mongodb"`
+	ID      string       `validate:"required,mongodb"`
+	Session *jwt.Session `validate:"-"`
 }
 
 func NewDeleteUserByIDCmd(repo repos.UsersRepo) *DeleteUserByIDCmd {
@@ -18,6 +20,11 @@ func NewDeleteUserByIDCmd(repo repos.UsersRepo) *DeleteUserByIDCmd {
 }
 
 func (c *DeleteUserByIDCmd) Exec(input *DeleteUserByIDCmdInput) error {
+	sess := input.Session
+	if sess == nil || (!sess.IsAdmin && input.ID != sess.UserID) {
+		return exception.NewForbiddenException()
+	}
+
 	err := c.repo.DeleteByID(input.ID)
 	if err != nil {
 		return exception.NewRepoException(
